+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+[unreleased]
+## Added
+- Added new unit test cases.
+
 [2.1.18] - 2025-02-19
 ### Fixed
 - Fixed security issue for tomcat TOCTOU