+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 [Unreleased]
+
+## [2.1.21] - 2025-03-10
 ### Added
 - New changes for code E2010 in service offering.